move the directory to get codecov test couverage
