+++ server/src/test/java/com/alibaba/otter/canal/server/embedded/CanalServerWithEmbedded_FileModeTest.java
@@ -1,42 +1,27 @@
-package com.alibaba.otter.canal.server;
+package com.alibaba.otter.canal.server.embedded;
 
 import java.net.InetSocketAddress;
 import java.util.Arrays;
 
-import org.I0Itec.zkclient.ZkClient;
-import org.junit.Before;
-
-import com.alibaba.otter.canal.common.zookeeper.ZookeeperPathUtils;
 import com.alibaba.otter.canal.instance.manager.model.Canal;
 import com.alibaba.otter.canal.instance.manager.model.CanalParameter;
-import com.alibaba.otter.canal.instance.manager.model.CanalParameter.HAMode;
-import com.alibaba.otter.canal.instance.manager.model.CanalParameter.IndexMode;
-import com.alibaba.otter.canal.instance.manager.model.CanalParameter.MetaMode;
-import com.alibaba.otter.canal.instance.manager.model.CanalParameter.SourcingType;
-import com.alibaba.otter.canal.instance.manager.model.CanalParameter.StorageMode;
+import com.alibaba.otter.canal.instance.manager.model.CanalParameter.*;
 
-public class CanalServerWithEmbedded_StandbyTest extends BaseCanalServerWithEmbededTest {
-
-    private ZkClient zkClient = new ZkClient(cluster1);
-
-    @Before
-    public void setUp() {
-        zkClient.deleteRecursive(ZookeeperPathUtils.CANAL_ROOT_NODE);
-        super.setUp();
-    }
+public class CanalServerWithEmbedded_FileModeTest extends BaseCanalServerWithEmbededTest {
 
     protected Canal buildCanal() {
         Canal canal = new Canal();
         canal.setId(1L);
         canal.setName(DESTINATION);
-        canal.setDesc("test");
+        canal.setDesc("my standalone server test ");
 
         CanalParameter parameter = new CanalParameter();
 
-        parameter.setZkClusters(Arrays.asList("127.0.0.1:2188"));
-        parameter.setMetaMode(MetaMode.MIXED); // 冷备，可选择混合模式
+        parameter.setMetaMode(MetaMode.LOCAL_FILE);
+        parameter.setDataDir("./conf");
+        parameter.setMetaFileFlushPeriod(1000);
         parameter.setHaMode(HAMode.HEARTBEAT);
-        parameter.setIndexMode(IndexMode.META);// 内存版store，需要选择meta做为index
+        parameter.setIndexMode(IndexMode.MEMORY_META_FAILBACK);
 
         parameter.setStorageMode(StorageMode.MEMORY);
         parameter.setMemoryStorageBufferSize(32 * 1024);
@@ -46,8 +31,8 @@ public class CanalServerWithEmbedded_StandbyTest extends BaseCanalServerWithEmbe
             new InetSocketAddress(MYSQL_ADDRESS, 3306)));
         parameter.setDbUsername(USERNAME);
         parameter.setDbPassword(PASSWORD);
-        parameter.setPositions(Arrays.asList("{\"journalName\":\"mysql-bin.000001\",\"position\":6163L,\"timestamp\":1322803601000L}",
-            "{\"journalName\":\"mysql-bin.000001\",\"position\":6163L,\"timestamp\":1322803601000L}"));
+        parameter.setPositions(Arrays.asList("{\"journalName\":\"mysql-bin.000001\",\"position\":332L,\"timestamp\":\"1505998863000\"}",
+            "{\"journalName\":\"mysql-bin.000001\",\"position\":332L,\"timestamp\":\"1505998863000\"}"));
 
         parameter.setSlaveId(1234L);
